rework how cells are filled
