+++ CoEpi/ui/home/HomeViewController.swift
@@ -10,7 +10,7 @@ class HomeViewController: UIViewController{
     
     private let disposeBag = DisposeBag()
 
-    private var dataSource = HomeListDataSource()
+    //private var dataSource = HomeListDataSource()
 
     
     init(viewModel: HomeViewModel) {
@@ -45,11 +45,22 @@ class HomeViewController: UIViewController{
         
         tableView.register(cellClass: UITableViewCell.self)
         tableView.tableFooterView = UIView()
+        tableView.rowHeight = 120.0
         
-        viewModel.homeEntries
-            .drive(tableView.rx.items(dataSource: dataSource))
+        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "Cell")
+
+        Observable.just(["Health Quiz", "Contact Alerts", (getVersionNumber() + " " + getBuildNumber())])
+            .bind(to: tableView.rx.items) { (tableView, row, element) in
+                let cell = tableView.dequeueReusableCell(withIdentifier: "Cell")!
+                cell.textLabel?.text = "\(element)"
+                cell.textLabel?.font = .systemFont(ofSize: 14)
+                cell.backgroundColor = .clear
+                cell.selectionStyle = UITableViewCell.SelectionStyle.none
+                cell.accessoryType = UITableViewCell.AccessoryType.disclosureIndicator
+                return cell
+            }
             .disposed(by: disposeBag)
-        
+
         tableView.rx.itemSelected
             .subscribe(onNext: {indexPath in
                 if indexPath[1] == 0{
@@ -69,35 +80,18 @@ class HomeViewController: UIViewController{
     }
 }
 
-private class HomeListDataSource: NSObject, RxTableViewDataSourceType {
-    private var homeEntries: [HomeEntryViewData] = []
-
-    func tableView(_ tableView: UITableView, observedEvent: RxSwift.Event<[HomeEntryViewData]>) {
-        if case let .next(homeEntries) = observedEvent {
-            self.homeEntries = homeEntries
-            tableView.reloadData()
-        }
+private func getVersionNumber() -> String{
+    guard let version = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String
+     else{
+        fatalError("Failed to read bundle version")
     }
+    print("Version : \(version)");
+    return "Version: \(version)"
 }
-
-extension HomeListDataSource: UITableViewDataSource {
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        homeEntries.count
+private func getBuildNumber() -> String {
+    guard let build = Bundle.main.infoDictionary?["CFBundleVersion"] as? String else {
+        fatalError("Failed to read build number")
     }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell: UITableViewCell = tableView.dequeue(cellClass: UITableViewCell.self, forIndexPath: indexPath)
-        let label = cell.textLabel
-        label?.font = .systemFont(ofSize: 14)
-
-        switch homeEntries[indexPath.row]{
-            case .Header(let text):
-                label?.text = text
-                cell.backgroundColor = .clear
-                cell.selectionStyle = UITableViewCell.SelectionStyle.none
-                cell.accessoryType = UITableViewCell.AccessoryType.disclosureIndicator
-                return cell
-        }
-    }
+    print("Build : \(build)")
+    return "Build: \(build)"
 }
